+++ lib/doorkeeper/oauth/pre_authorization.rb
@@ -48,7 +48,11 @@ module Doorkeeper
 
       def validate_scopes
         return true unless scope.present?
-        Helpers::ScopeChecker.valid? scope, server.scopes
+        if client.application.scopes.empty?
+          Helpers::ScopeChecker.valid?(scope, server.scopes)
+        else
+          Helpers::ScopeChecker.valid?(scope, server.scopes & client.application.scopes)
+        end
       end
 
       # TODO: test uri should be matched against the client's one