+++ lib/doorkeeper/orm/mongoid4/application.rb
@@ -2,6 +2,7 @@ module Doorkeeper
   class Application
     include Mongoid::Document
     include Mongoid::Timestamps
+    include Models::Mongoid4::Scopes
 
     include ApplicationMixin